I'm going to implement the post-mortem feature and refactor the context.
