+++ src/context/AuthContext.tsx
@@ -1,0 +1,85 @@
+import React, { createContext, useState, useEffect, useContext } from 'react';
+import {
+    onAuthChange,
+    anonymousSignIn,
+    customTokenSignIn,
+    onUserSnapshot,
+    onPartnerSnapshot,
+    linkPartners,
+    createUserProfile,
+    updateUserName as updateUserNameInDb,
+} from '../services/firebase';
+
+export const AuthContext = createContext(null);
+
+export const AuthProvider = ({ children }) => {
+    const [user, setUser] = useState(null);
+    const [partner, setPartner] = useState(null);
+    const [isLoading, setIsLoading] = useState(true);
+    const [notification, setNotification] = useState({ show: false, message: '', type: 'info', duration: 4000 });
+
+
+    useEffect(() => {
+        const unsubscribe = onAuthChange(async (currentUser) => {
+            if (currentUser) {
+                onUserSnapshot(currentUser.uid, async (snap) => {
+                    if (snap.exists()) {
+                        const userData = snap.data();
+                        setUser({ uid: currentUser.uid, ...userData });
+                        if (userData.partnerId) {
+                            onPartnerSnapshot(userData.partnerId, (partnerSnap) => {
+                                if (partnerSnap.exists()) {
+                                    setPartner({ uid: userData.partnerId, ...partnerSnap.data() });
+                                } else {
+                                    setPartner(null);
+                                }
+                            });
+                        } else {
+                            setPartner(null);
+                        }
+                    } else {
+                        const urlParams = new URLSearchParams(window.location.search);
+                        const inviterId = urlParams.get('invite');
+                        if (inviterId && inviterId !== currentUser.uid) {
+                            await linkPartners(inviterId, currentUser.uid);
+                            window.history.replaceState({}, document.title, window.location.pathname);
+                        } else {
+                            await createUserProfile(currentUser.uid);
+                        }
+                    }
+                });
+            } else {
+                try {
+                    const token = typeof __initial_auth_token !== 'undefined' ? __initial_auth_token : null;
+                    if (token) {
+                        await customTokenSignIn(token);
+                    } else {
+                        await anonymousSignIn();
+                    }
+                } catch (error) {
+                    console.error("Authentication Error:", error);
+                    setNotification({show: true, message: "Authentication failed. Please refresh.", type: 'warning'});
+                }
+            }
+            setIsLoading(false);
+        });
+        return () => unsubscribe();
+    }, []);
+
+    const value = {
+        user,
+        partner,
+        isLoading,
+        updateUserName: (newName) => updateUserNameInDb(user.uid, newName),
+    };
+
+    return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
+};
+
+export const useAuth = () => {
+    const context = useContext(AuthContext);
+    if (context === null) {
+        throw new Error('useAuth must be used within an AuthProvider');
+    }
+    return context;
+}
