--- src/hooks/useAppContext.ts
@@ -1,10 +1,0 @@
-import { useContext } from 'react';
-import { AppContext } from '../context/AppContext';
-
-export const useAppContext = () => {
-    const context = useContext(AppContext);
-    if (!context) {
-        throw new Error('useAppContext must be used within an AppProvider');
-    }
-    return context;
-};
